requirements.txt
##############################################################################
# AI-RIDS — Python Dependencies
# Tested on Python 3.9+
##############################################################################

# ── Core ML / Data ────────────────────────────────────────────────────────────
lightgbm>=4.3.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.0
scipy>=1.12.0

# ── Telemetry / System ────────────────────────────────────────────────────────
psutil>=5.9.8             # Cross-platform CPU/memory/process metrics
watchdog>=4.0.0           # File-system event monitoring
scapy>=2.5.0              # Packet capture and analysis
pywin32>=306              # Windows API bindings (PDH, VSS, WMI)
wmi>=1.5.1                # Windows WMI (alternative HPC path)
pyshark>=0.6              # Optional: Wireshark/tshark-based capture

# ── UI/API bridge ─────────────────────────────────────────────────────────────
flask>=3.0.0              # REST API bridge for the React dashboard
flask-cors>=4.0.0         # CORS support for Flask

# ── Configuration / Serialisation ────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ── Logging & Metrics ─────────────────────────────────────────────────────────
colorlog>=6.8.0
rich>=13.7.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0

# ── Utility ───────────────────────────────────────────────────────────────────
click>=8.1.7              # CLI interface for main.py
tabulate>=0.9.0           # Pretty-print threat tables
